Advections

Advection functions

SemiLagrangian.interpolate — Method.

interpolate( coeffs, n, f)

Compute interpolatted value at x from coefficients get from compute_interpolants.

This function is a julia version of a Fortran code written by

Edwin Chacon Golcher (Institute of Physics of the Czech Academy of Sciences)

source

SemiLagrangian.interpolate — Method.

interpolate( p, f, delta, alpha)

Compute the interpolating spline of degree p of odd degree of a 1D function f on a periodic uniform mesh, at all points x-alpha input f is the Fourier transform and complex you have to in2erse transform after return

source

SemiLagrangian.interpolate — Method.

interpolate( p, f, delta, alpha)

Compute the interpolating spline of degree p of odd degree of a 1D function f on a periodic uniform mesh, at all points x-alpha. f type is Vector{Float64}.

Derived from a Python program written by Eric Sonnendrucker (Max-Planck-Institut fur Plasmaphysik - Garching)

source

SemiLagrangian.advection! — Method.

advection!(f, mesh, v, n2, dt, interp)

Advection of a 2d function f along its first dimension with velocity v. Since the fft are computed inplace, the function must be represented by a Array{Complex{Float64},2}.

source

SemiLagrangian.advection! — Method.

 advection!( f, mesh1, v,  dt)

Semi-lagrangian advection function of 2D distribution function represented by array f. The advection operates along axis (=1 is most efficient) with speed v during dt.

It uses cubic splines interpolation.

source

SemiLagrangian.advection! — Method.

advection!( mesh, f, v, dt, interp, axis)

Advection of a 2d function f discretized on a 2d mesh along the input axis at velocity v

source

SemiLagrangian.bspline — Method.

bspline(p, j, x)

Return the value at x in [0,1[ of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the de Boor's recursion formula

Derived from a Python program written by Eric Sonnendrucker (Max-Planck-Institut fur Plasmaphysik - Garching))

source

SemiLagrangian.compute_interpolants — Method.

compute_interpolants( n, f)

Compute interpolation coefficients

This function is a julia version of a Fortran code written by Edwin Chacon Golcher (Institute of Physics of the Czech Academy of Sciences)

source

SemiLagrangian.interpolate! — Method.

interpolation par spline periodique dans les deux directions. Les points d'interpolation sont definis grace a depx et depy qui definissent le deplacement par rapport au maillage.

f contient les valeurs de la fonction de distribution

depx et depy : deplacements par rapport au maillage des points dans les quels on veut evaluer la spline.

source

SemiLagrangian.nat_x! — Method.

natural splines

source