Mesh

SemiLagrangian.UniformMesh — Type

UniformMesh{T}
UniformMesh(start::T, stop::T, length::Int) where {T}

1D uniform mesh data.

Arguments

start::T : beginning of the mesh

stop::T : end of the mesh

length::Int : number of cells of the mesh

Implementation

step::T : size step

points::Vector{T}: Array with node positions

width::T : Distance between left and right edges.

Base.step — Function

Base.step(mesh::UniformMesh)

Get the step of the mesh

Argument

mesh::UniformMesh : the mesh

Return

step : the step of the mesh that is the difference between two contiguous points

Base.length — Function

Base.length(mesh::UniformMesh)

Get the length of the mesh

Argument

mesh::UniformMesh : the mesh

Return

length : the length of the mesh that is the number of points and cells

SemiLagrangian.points — Function

points(mesh::UniformMesh)

Get the points of the mesh

Argument

mesh::UniformMesh : the mesh

Return

points : the points of the mesh that is the vector of all points of the mesh except the last

SemiLagrangian.width — Function

width(mesh::UniformMesh)

Get the width of the mesh

Argument

mesh::UniformMesh : the mesh

Return

width : the width that is step*length or distance between left and right edges.

SemiLagrangian.vec_k_fft — Function

vec_k_fft(mesh::UniformMesh{T}) where{T}

Get the fft coefficients of the mesh

Argument

mesh::UniformMesh{T} : the mesh

Return

fft coefficients

Interpolation