Functions

Functions and types

SemiLagrangian.Advection — Type.

Advection(interpolation_type, mesh, LBC, RBC)

Creates a 1d backward semi-lagrangian advection.

interp : Interpolation type (Bspline(degree), Lagrange(degree))

mesh : UniformMesh along advection direction

LBC, RBC : Boundary conditions type (:periodic, :Hermite)

source

SemiLagrangian.Advection — Method.

advection!(f, v, dt)

Advection of a 2d function f discretized on a 2d mesh along the input axis at velocity v. This function is created from Advector callable type.

```julia mesh = UniformMesh( -π, π, 64 ) advection! = Advection( mesh, Bspline(3), :periodic )

f = exp.( - mesh.points .^ 2 )

dt = 0.5 v = ones( Float64, mesh.length)

advection!( f, v, dt )

source

SemiLagrangian.PeriodicAdvection — Type.

advection! = PeriodicAdvection( mesh, bspl )

Type to perform 1d advection on periodic domain. Bspline interpolation is used combined with fft transform to solve the system. Bspline degree must be odd.

p = 5
nx = 128

x1min, x1max = -10, 10

mesh1 = UniformMesh(x1min, x1max, n1; endpoint=false)

advection! = PeriodicAdvection( mesh1, Bspline(p) )

advection! computes the interpolating spline of degree p of odd degree of first dimension of array f on a periodic uniform mesh, at all points x-alpha. f type is Array{Float64,2}.

source

SemiLagrangian.UniformMesh — Type.

UniformMesh(start, stop, length)

1D uniform mesh data.

length : number of points length-1 : number of cells

If you want remove the last point for periodic domain, set endpoint=false

- `step`  : size step
- `points`: Array with node positions
- `width` : Distance between left and right edges.

source

SemiLagrangian.compute_charge! — Method.

compute_rho!( rho, mesh_v, fvx)

Compute charge density

ρ(x,t) = ∫ f(x,v,t) dv

source

SemiLagrangian.compute_e! — Method.

compute_e!( e, mesh, ρ)

∇.e = - ρ

Inplace computation of electric field. Fields e and rho are already allocated.

source

SemiLagrangian.bspline — Method.

bspline(p, j, x)

Return the value at x in [0,1[ of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the de Boor's recursion formula using the De Boor's Algorithm

\[B_{i,0}(x) := \left\{ \begin{matrix} 1 & \mathrm{if} \quad t_i ≤ x < t_{i+1} \\ 0 & \mathrm{otherwise} \end{matrix} \right.\]

\[B_{i,p}(x) := \frac{x - t_i}{t_{i+p} - t_i} B_{i,p-1}(x) + \frac{t_{i+p+1} - x}{t_{i+p+1} - t_{i+1}} B_{i+1,p-1}(x).\]

source

SemiLagrangian.eval_basis! — Method.

eval_basis!( spl, x, values )

Evaluate value at x of all basis functions with support in local cell values[j] = B_j(x) for jmin <= j <= jmin+degree

source

SemiLagrangian.eval_deriv! — Method.

eval_deriv!( derivs, spl, x )

Evaluate derivative at x of all basis functions with support in local cell derivs[j] = B_j'(x) for jmin <= j <= jmin+degree

source

SemiLagrangian.eval_deriv — Method.

eval_deriv( spl, x )

Evaluate derivative of 1D spline at location x: y=S'(x)

source

SemiLagrangian.eval_value — Method.

eval_value( spl, x )

Evaluate value of 1D spline at location x: y=S(x)

source

SemiLagrangian.interpolate! — Method.

interpolation par spline periodique dans les deux directions. Les points d'interpolation sont definis grace a depx et depy qui definissent le deplacement par rapport au maillage.

f contient les valeurs de la fonction de distribution

depx et depy : deplacements par rapport au maillage des points dans les quels on veut evaluer la spline.

source

SemiLagrangian.nat_x! — Method.

natural splines

source