Mesh

SemiLagrangian.UniformMesh — Type

UniformMesh{T}
UniformMesh(start::T, stop::T, length::Int) where {T}

1D uniform mesh data.

Arguments

start::T : beginning of the mesh

stop::T : end of the mesh

length::Int : number of cells of the mesh

Implementation

step::T : size step

points::Vector{T}: Array with node positions

width::T : Distance between left and right edges.

Base.step — Function

Base.step(mesh::UniformMesh)

Get the step of the mesh

Argument

mesh::UniformMesh : the mesh

Return

step : the step of the mesh that is the difference between two contiguous points

Base.length — Function

Base.length(mesh::UniformMesh)

Get the length of the mesh

Argument

mesh::UniformMesh : the mesh

Return

length : the length of the mesh that is the number of points and cells

SemiLagrangian.points — Function

points(mesh::UniformMesh)

Get the points of the mesh

Argument

mesh::UniformMesh : the mesh

Return

points : the points of the mesh that is the vector of all points of the mesh except the last

SemiLagrangian.width — Function

width(mesh::UniformMesh)

Get the width of the mesh

Argument

mesh::UniformMesh : the mesh

Return

width : the width that is step*length or distance between left and right edges.

SemiLagrangian.vec_k_fft — Function

vec_k_fft(mesh::UniformMesh{T}) where{T}

Get the fft coefficients of the mesh

Argument

mesh::UniformMesh{T} : the mesh

Return

fft coefficients

Interpolation

SemiLagrangian.AbstractInterpolation — Type

AbstractInterpolation{T, edge, order}

Abstract supertype for all interpolation type

Implementation constraint

tabfct::Vector : this attribut must be on the implementation, it is a table of function of size order+1

SemiLagrangian.get_order — Function

get_order(_::AbstractInterpolation{T, edge, order}) where{T, edge, order}

Return the order of interpolation implementation

SemiLagrangian.sol — Method

sol(_::AbstractInterpolation, line::AbstractVector)

Interface method to transform the treated line, by default this method does nothing

Arguments :

_::AbstractInterpolation : interpolation implementation

line::AbstractVector : line to transform

Return :

The transformed line

SemiLagrangian.interpolate! — Function

interpolate!( fp::AbstractVector{T}, 
    fi::AbstractVector{T},
    decint::Int, 
    precal::Vector{T}, 
    interp::AbstractInterpolation{T, CircEdge, order},
    tabmod=gettabmod(length(fi)) ) where {T, order}

apply an offset to the function fi interpolate by interp struct, the result is in fp vector, decint and precal are precompute with get_precal method, the TypeEdge is CircEdge

Arguments

fp::AbstractVector : output vector

fi::AbstractVector : input vector

decint : offset in units of dx

precal::Vector : vector of length order+1 precompute with get_precal(interp, dec) (dec is the offset)

interp::AbstractInterpolation{T, CircEdge, order} : interpolation implementation, note that TypeEdge is CircEdge

tabmod=gettabmod(length(fi)) : precompute for "begin at one" modulo

Returns :

No return

interpolate!( 
fp::AbstractVector{T}, fi::AbstractVector{T}, decint::Int, 
allprecal::Vector{Vector{T}}, 
interp::AbstractInterpolation{T, InsideEdge, order},
tabmod=gettabmod(length(fi))
) where {T, order}

apply an offset to the function fi interpolate by interp struct, the result is in fp vector, decint and precal are precompute with get_precal method, the TypeEdge is InsideEdge, it is a marginal case

Arguments

fp::AbstractVector : output vector

fi::AbstractVector : input vector

decint : offset in units of dx

allprecal::Vector{Vector{T}} : vector of vector of length order+1 precompute with get_precal(interp, dec) (dec is the offset)

interp::AbstractInterpolation{T, InsideEdge, order} : interpolation implementation, note that TypeEdge is CircEdge

tabmod=gettabmod(length(fi)) : precompute for "begin at one" modulo

Returns :

No return

interpolate!( fp, fi, dec, interp)

apply the offset dec to the function fi interpolate by interp struct, the result is in fp Vector

Arguments

fp : output vector of length n

fi : input vector of length n

dec : offset in units of dx

interp::AbstractInterpolation : interpolation implementation

Returns :

No return