Functions

Functions and types

SemiLagrangian.Advection — Type.

Advection(interpolation_type, mesh, LBC, RBC)

Creates a 1d backward semi-lagrangian advection.

interp : Interpolation type (Bspline(degree), Lagrange(degree))

mesh : UniformMesh along advection direction

LBC, RBC : Boundary conditions type (:periodic, :Hermite)

source

SemiLagrangian.Advection — Method.

advection!(f, v, dt)

Advection of a 2d function f discretized on a 2d mesh along the input axis at velocity v. This function is created from Advector callable type.

```julia mesh = UniformMesh( -π, π, 64 ) advection! = Advection( Bspline(3), mesh, 1, :periodic, :periodic )

f = exp.( - mesh.points .^ 2 )

dt = 0.5 v = ones( Float64, mesh.length)

advection!( f, v, dt )

source

SemiLagrangian.bspline — Method.

bspline(p, j, x)

Return the value at x in [0,1[ of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the de Boor's recursion formula

Original python program written by Eric Sonnendrucker (Max-Planck-Institut fur Plasmaphysik - Garching (Germany))

source

SemiLagrangian.eval_basis! — Method.

Evaluate value at x of all basis functions with support in local cell values[j] = B_j(x) for jmin <= j <= jmin+degree

source

SemiLagrangian.eval_deriv! — Method.

Evaluate derivative at x of all basis functions with support in local cell derivs[j] = B_j'(x) for jmin <= j <= jmin+degree

source

SemiLagrangian.eval_deriv — Method.

Evaluate derivative of 1D spline at location x: y=S'(x)

source

SemiLagrangian.eval_value — Method.

Evaluate value of 1D spline at location x: y=S(x)

source

SemiLagrangian.interpolate! — Method.

interpolate!( f, bspl, alpha)

Compute the interpolating spline of degree p of odd degree of a 1D function f on a periodic uniform mesh, at all points x-alpha

Array representing f is complex and modified inplace.

source

SemiLagrangian.interpolate! — Method.

interpolation par spline periodique dans les deux directions. Les points d'interpolation sont definis grace a depx et depy qui definissent le deplacement par rapport au maillage.

f contient les valeurs de la fonction de distribution

depx et depy : deplacements par rapport au maillage des points dans les quels on veut evaluer la spline.

source

SemiLagrangian.interpolate — Method.

interpolate( f, bspl, alpha)

Compute the interpolating spline of degree p of odd degree of a 1D function f on a periodic uniform mesh, at all points x-alpha. f type is Vector{Float64}.

Derived from a Python program written by Eric Sonnendrucker (Max-Planck-Institut fur Plasmaphysik - Garching)

source

SemiLagrangian.nat_x! — Method.

natural splines

source