Functions and types

SemiLagrangian.Advection — Type

Advection{T, Nsp, Nv, Nsum, timeopt}
Advection(
    t_mesh_sp::NTuple{Nsp, UniformMesh{T}},
    t_mesh_v::NTuple{Nv, UniformMesh{T}},
    t_interp_sp::NTuple{Nsp, InterpolationType{T}},
    t_interp_v::NTuple{Nv, InterpolationType{T}},
    dt_base::T;
    tab_coef=[1//2, 1//1, 1//2],
    tab_fct=[identity, identity, identity],
    timeopt::TimeOptimization=NoTimeOpt
)

Immutable structure that contains constant parameters for multidimensional advection

Type parameters

T::DataType : type of data

Nsp : number of space dimensions

Nv : number of velocity dimensions

Nsum : the total number of dimensions (Nsum = Nsp + Nv)

timeopt::TimeOptimization : time optimization

Arguments

t_mesh_sp::NTuple{Nsp, UniformMesh{T}} : tuple of space meshes (one per space dimension)

t_mesh_v::NTuple{Nv, UniformMesh{T}} : tuple of velocity meshes (one per velocity dimension)

t_interp_sp::NTuple{Nsp, InterpolationType{T}} : tuple of space interpolations (one per space dimension)

t_interp_v::NTuple{Nv, InterpolationType{T}} : tuple of velocity interpolations(one per velocity dimension)

dt_base::T : time delta for one advection series

Keywords

tab_coef=[1//2, 1//1, 1//2] : coefficient table for one advection series, the coefficients at odd indexes is for space advection series, the coefficients at even indexes is for velocity advection series

tab_fct=[identity, identity, identity] : function table for one advection series, with the same indexes than tab_coef

Implementation

sizeall : tuple of the sizes of all dimensions (space before velocity)

sizeitr : tuple of iterators of indexes of each dimension

t_mesh_sp : tuple of space meshes

t_mesh_v : tuple of velocity meshes

t_interp_sp : tuple of space interpolation types

t_interp_v : tuple of velocity interpolation types

dt_base::T : time unit of an advection series

tab_coef : coefficient table

tab_fct : function table

v_square : precompute for ke

nbsplit : number of slices for split

mpiid : MPI id

Throws

ArgumentError : Nsp must be less or equal to Nv.

SemiLagrangian.AdvectionData — Type

AdvectionData{T,Nsp,Nv,Nsum,timeopt}
AdvectionData(
adv::Advection{T,Nsp,Nv,Nsum,timeopt}, 
data::Array{T,Nsum},
parext)

Mutable structure that contains variable parameters of advection series

Type parameters

T::DataType : type of data

Nsp : number of space dimensions

Nv : number of velocity dimensions

Nsum : the total number of dimensions (Nsum = Nsp + Nv)

timeopt::TimeOptimization : time optimization

Arguments

adv::Advection{T,Nsp,Nv,Nsum} : link to the constant data of this advection

data::Array{T,Nsum} : Initial data of this advection

parext : external data of this advection to compute alpha of each interpolations

Implementation

adv::Advection{T,Nsp,Nv,Nsum,timeopt} : link to the constant data of this advection

state_coef::Int : state that is the index of tab_coef, it is from one to lenth(tab_coef)

state_dim::Int : the dimension index, from 1 to Nsp in space states, from one to Nv in velocity state

data:Array{T,Nsum} : it is the working buffer

bufdata::Vector{T} : vector of the same size of the working buffer

t_buf::NTuple{Nsum, Array{T,2}} : tuple of buffer that is used to get the linear data for interpolation, one buffer per thread

cache_alpha::Union{T,Nothing} : cache for precal, the precal is compute only when the alpha or decint values change

cache_decint::Int64 : for precal cache

cache_precal::Vector{T} : for precal cache

parext::ExtDataAdv : external data of this advection to compute alpha of each interpolations

Methods to define

initcoef!(parext::AbstractExtDataAdv, self::AdvectionData) : this method called at the beginning of each advection to initialize parext data. The self.parext mutable structure is the only data that init! can modify otherwise it leads to unpredictable behaviour.

getalpha(parext::AbstractExtDataAdv, self::AdvectionData, ind) : return the alpha number that is used for interpolation.

getperm(parext::AbstractExtDataAdv, advd::AdvectionData) : get the permutation of the dimension as a function of the current state

SemiLagrangian.PoissonConst — Type

PoissonConst{T, Nsp, Nv}
PoissonConst(adv::Advection{T, Nsp, Nv, Nsum}; isfftbig=true)

Constant data for the computation of poisson coefficients

Arguments

adv::Advection{T, Nsp, Nv, Nsum, timeopt} : Advection constant data

isfftbig=true: if true compute the fttbig structure

Implementation

adv : Advection constant data

`v_k' : vector of vector of fourier coefficents for integration for each space dimension

fctv_k : Array of space dimensions of the inverse of the norm of fourier coefficients

pfftbig : Fourier data for space dimensions

SemiLagrangian.PoissonVar — Type

PoissonVar{T, Nsp, Nv} <: AbstractExtDataAdv
PoissonVar(pc::PoissonConst{T, Nsp, Nv})

mutable structure of variable data for the poisson computation

Arguments

pc::PoissonConst{T, Nsp, Nv} : poisson constant data

Implementation

pc::PoissonConst{T, Nsp, Nv} : poisson constant data

rho::Array{T, Nsp} : result of the compute_charge that is the sum along velocity dimensions

t_elfield::NTuple{Nsp,Array{Complex{T}, Nsp}} : electric fields initialized at each beginning of velocity advection subseries

SemiLagrangian.UniformMesh — Type

UniformMesh(start, stop, length)

1D uniform mesh data.

length : number of points length-1 : number of cells

If you want remove the last point for periodic domain, set endpoint=false

Arguments

- `step`  : size step
- `points`: Array with node positions
- `width` : Distance between left and right edges.

SemiLagrangian._getpolylagrange — Method

_getpolylagrange(k::Int64, order::Int64, origin::Int64, N::DataType)

Function that return the k-th Lagrange Polynomial of a certain order. Coefficients are rational then the return is exact. The polynomial is equal to : $\prod_{i=0,\ i \neq k}^{order} \frac{x - i - origin}{k - i}$

Arguments

k::Int64 : number of the Polynomial, k must be between 0 and order (0<= k <= order).

order::Int64 : order of the polynomial.

origin::Int64 : origin of the first indice.

N::DataType : type of the Integer that is the base of the rational type, in fact Int64 or BigInt. BigInt is needed for order greater or equal to 21.

Returns

Polynomial{Rational{N}} : the k-th Lagrange polynomial of order order

Throws

DommaineError : when 0 <= k <= order is false or when N ∉ {BInt64, BigInt}

SemiLagrangian.advection! — Method

advection!(self::AdvectionData)

Advection function of a multidimensional function f discretized on mesh

Argument

self::AdvectionData : mutable structure of variables data

Return value

true : means that the advection series must continue

false : means that the advection series is ended.

SemiLagrangian.compute_charge! — Method

compute_charge!( self::AdvectionData)

Compute charge density

ρ(x,t) = ∫ f(x,v,t) dv

Argument

self::AdvectionData : mutable structure of variables data.

SemiLagrangian.compute_charge! — Method

compute_charge!( rho, mesh_v, fvx)

Compute charge density

ρ(x,t) = ∫ f(x,v,t) dv

Arguments

rho::Array{T,Nsp} : result table correctly sized, it is he output

t_mesh_v::NTuple{Nv,UniformMesh{T}} : velocity meshes

f::Array{T,Nsum} : input data

SemiLagrangian.compute_ee — Method

compute_ee(self::AdvectionData)

compute electric enegie || E(t,.) ||_L2

Argument

self::AdvectionData : veriable advection data structure.

SemiLagrangian.compute_ee — Method

compute_ee(t_mesh_sp, t_elf)

compute electric enegie || E(t,.) ||_L2

Arguments

t_mesh_sp::NTuple{N,UniformMesh{T}} : tuple of space meshes

t_elf::NTuple{N,Array{T,N}} : tuple of electric field

SemiLagrangian.compute_elfield! — Method

compute_elfield!( self:AdvectionData)

computation of electric field ∇.e = - ρ

Argument

self::AdvectionData : mutable structure of variables data.

SemiLagrangian.compute_ke — Method

compute_ke(self::AdvectionData)

kinetic Energie

1/2∫∫ v^2 f(x,v,t) dv dx

Arguments

self::AdvectionData : mutable structure of variables data.

SemiLagrangian.compute_ke — Method

compute_ke(t_mesh_sp, t_mesh_v, f)

kinetic Energie

1/2∫∫ v^2 f(x,v,t) dv dx

Arguments

t_mesh_sp::NTuple{Nsp, UniformMesh{T}} : tuple of space meshes

t_mesh_v::NTuple{Nv, UniformMesh{T}} : tuple of velocity meshes

f::Array{T,Nsum} : function data.

SemiLagrangian.getalpha — Method

getalpha(self::AdvectionData{T, Nsp, Nv, Nsum}, ind)

Implementation of the interface function that is called before each interpolation in advection

SemiLagrangian.interpolate! — Method

interpolate!( fp, fi, dec, interp)

apply the offset dec to the function fi interpolate by interp struct, the result is in fp Vector

Arguments

fp : output vector of length n

fi : input vector of length n

dec : offset in units of dx

interp::InterpolationType : interpolation implementation

Returns :

No return

SemiLagrangian.interpolate! — Method

interpolate!( fp, fi, decint, precal, interp)

apply an offset to the function fi interpolate by interp struct, the result is in fp vector, decint and precal are precompute with get_precal method.

Arguments

fp::AbstractVector : output vector

fi::AbstractVector : input vector

decint : offset in units of dx

precal::Vector : vector of length order+1 precompute with get_precal(interp, dec) (dec is the offset)

interp::InterpolationType : interpolation implementation

Returns :

No return

SemiLagrangian.nextstate! — Method

nextstate!(self::AdvectionData{T, Nsp, Nv, Nsum})

Function called at the end of advection function to update internal state of AdvectionData structure

Argument

self::AdvectionData{T, Nsp, Nv, Nsum} : object to update

return value

ret::Bool : true if the series must continue false at the end of the series.