Functions

Base.length — Method

Base.length(mesh::UniformMesh)

Get the length of the mesh

Argument

mesh::UniformMesh : the mesh

Return

length : the length of the mesh that is the number of points and cells

Base.step — Method

Base.step(mesh::UniformMesh)

Get the step of the mesh

Argument

mesh::UniformMesh : the mesh

Return

step : the step of the mesh that is the difference between two contiguous points

SemiLagrangian._getpolylagrange — Method

_getpolylagrange(k::Int64, order::Int64, origin::Int64)

Function that return the k-th Lagrange Polynomial of a certain order. Coefficients are rational then the return is exact. The polynomial is equal to : $\prod_{i=0,\ i \neq k}^{order} \frac{x - i - origin}{k - i}$

Arguments

k::Int64 : number of the Polynomial, k must be between 0 and order (0<= k <= order).

order::Int64 : order of the polynomial.

origin::Int64 : origin of the first indice.

Returns

Polynomial{Rational{BigInt}} : the k-th Lagrange polynomial of order order

Throws

DommaineError : when 0 <= k <= order is false or when N ∉ {BInt64, BigInt}

SemiLagrangian.advection! — Method

advection!(self::AdvectionData)

Advection function of a multidimensional function f discretized on mesh

Argument

self::AdvectionData : mutable structure of variables data

Return value

true : means that the advection series must continue

false : means that the advection series is ended.

SemiLagrangian.compute_charge! — Method

compute_charge!( self::AdvectionData)

Compute charge density

ρ(x,t) = ∫ f(x,v,t) dv

Argument

self::AdvectionData : mutable structure of variables data.

SemiLagrangian.compute_charge! — Method

compute_charge!(rho, mesh_v, f)

Compute charge density from phase space distribution f.

ρ(x,t) = ∫ f(x,v,t) dv

Arguments

rho::Array{T,Nsp}: output result density array.

t_mesh_v::NTuple{Nv,UniformMesh{T}}: velocity mesh.

f::Array{T,Nsum}: distribution function array.

SemiLagrangian.compute_ee — Method

compute_ee(self::AdvectionData)

Compute electric energy || E(t,.) ||_L2

Argument

self::AdvectionData: advection data structure.

SemiLagrangian.compute_ee — Method

compute_ee(t_mesh_sp, t_elf)

Compute electric energy || E(t,.) ||_L2

Arguments

t_mesh_sp::NTuple{N,UniformMesh{T}}: space mesh.

t_elf::NTuple{N,Array{T,N}}: electric field.

SemiLagrangian.compute_elfield! — Method

compute_elfield!( self:AdvectionData)

computation of electric field ∇.e = - ρ

Argument

self::AdvectionData : mutable structure of variables data.

SemiLagrangian.compute_ke — Method

compute_ke(self::AdvectionData)

Compute kinetic Energy.

∫∫ v^2 f(x,v,t) dv dx

Arguments

self::AdvectionData: mutable structure of variables data.

SemiLagrangian.compute_ke — Method

compute_ke(t_mesh_sp, t_mesh_v, f)

Compute kinetic Energy from phase space distribution f.

∫∫ v^2 f(x,v,t) dv dx

Arguments

t_mesh_sp::NTuple{Nsp, UniformMesh{T}}: space mesh.

t_mesh_v::NTuple{Nv, UniformMesh{T}}: velocity mesh.

f::Array{T,Nsum}: distribution function array.

SemiLagrangian.get_order — Method

get_order(_::AbstractInterpolation{T, edge, order}) where{T, edge, order}

Return the order of interpolation implementation

SemiLagrangian.getalpha — Method

getalpha(pv::RotationVar, self::AdvectionData, ind)

Implementation of the interface function that is called before each interpolation in advection

SemiLagrangian.getalpha — Method

getalpha(pv::TranslationVar, self::AdvectionData, i, ind)

Implementation of the interface function that is called before each interpolation in advection

SemiLagrangian.interpolate! — Method

interpolate!( 
fp::AbstractVector{T}, fi::AbstractVector{T}, decint::Int, 
allprecal::Vector{Vector{T}}, 
interp::AbstractInterpolation{T, InsideEdge, order},
tabmod=gettabmod(length(fi))
) where {T, order}

apply an offset to the function fi interpolate by interp struct, the result is in fp vector, decint and precal are precompute with getprecal method, the TypeEdge is InsideEdge, it is a marginal case

Arguments

fp::AbstractVector : output vector

fi::AbstractVector : input vector

decint : offset in units of dx

allprecal::Vector{Vector{T}} : vector of vector of length order+1 precompute with getprecal(interp, dec) (dec is the offset)

interp::AbstractInterpolation{T, InsideEdge, order} : interpolation implementation, note that TypeEdge is CircEdge

tabmod=gettabmod(length(fi)) : precompute for "begin at one" modulo

Returns :

No return

SemiLagrangian.interpolate! — Method

interpolate!( fp::AbstractVector{T}, 
    fi::AbstractVector{T},
    decint::Int, 
    precal::Vector{T}, 
    interp::AbstractInterpolation{T, CircEdge, order},
    self::AdvectionData{T},
    tabmod=gettabmod(size(fi)) ) where {T, order}

apply an offset to the function fi interpolate by interp struct, the result is in fp vector, decint and precal are precompute with getprecal method, the TypeEdge is CircEdge

Arguments

fp::AbstractVector : output vector

fi::AbstractVector : input vector

decint : offset in units of dx

precal::Vector : vector of length order+1 precompute with getprecal(interp, dec) (dec is the offset)

interp::AbstractInterpolation{T, CircEdge, order} : interpolation implementation, note that TypeEdge is CircEdge

tabmod=gettabmod(length(fi)) : precompute for "begin at one" modulo

Returns :

No return

SemiLagrangian.interpolate! — Method

interpolate!( fp, fi, dec, interp)

apply the offset dec to the function fi interpolate by interp struct, the result is in fp Vector

Arguments

fp : output vector of length n

fi : input vector of length n

dec : offset in units of dx

interp::AbstractInterpolation : interpolation implementation

Returns :

No return

SemiLagrangian.points — Method

points(mesh::UniformMesh)

Get the points of the mesh

Argument

mesh::UniformMesh : the mesh

Return

points : the points of the mesh that is the vector of all points of the mesh except the last

SemiLagrangian.retns — Method

nextstate!(self::AdvectionData{T, N})

Function called at the end of advection function to update internal state of AdvectionData structure

Argument

self::AdvectionData{T, N} : object to update

return value

ret::Bool : true if the series must continue false at the end of the series.

SemiLagrangian.sizeall — Method

sizeall(adv::Advection)

Return a tuple of the sizes of each dimensions

Argument

adv::Advection : Advection structure.

SemiLagrangian.sol — Method

sol(_::AbstractInterpolation, line::AbstractVector)

Interface method to transform the treated line, by default this method does nothing

Arguments :

_::AbstractInterpolation : interpolation implementation

line::AbstractVector : line to transform

Return :

The transformed line

SemiLagrangian.vec_k_fft — Method

vec_k_fft(mesh::UniformMesh{T}) where{T}

Get the fft coefficients of the mesh

Argument

mesh::UniformMesh{T} : the mesh

Return

fft coefficients

SemiLagrangian.width — Method

width(mesh::UniformMesh)

Get the width of the mesh

Argument

mesh::UniformMesh : the mesh

Return

width : the width that is step*length or distance between left and right edges.