Types

SemiLagrangian.Advection — Type

Advection{T, Nsp, Nv, Nsum, timeopt}
Advection(
    t_mesh_sp::NTuple{Nsp, UniformMesh{T}},
    t_mesh_v::NTuple{Nv, UniformMesh{T}},
    t_interp_sp::NTuple{Nsp, InterpolationType{T}},
    t_interp_v::NTuple{Nv, InterpolationType{T}},
    dt_base::T;
    tab_coef=[1//2, 1//1, 1//2],
    tab_fct=[identity, identity, identity],
    timeopt::TimeOptimization=NoTimeOpt
)

Immutable structure that contains constant parameters for multidimensional advection

Type parameters

T::DataType : type of data

Nsp : number of space dimensions

Nv : number of velocity dimensions

Nsum : the total number of dimensions (Nsum = Nsp + Nv)

timeopt::TimeOptimization : time optimization

Arguments

t_mesh_sp::NTuple{Nsp, UniformMesh{T}} : tuple of space meshes (one per space dimension)

t_mesh_v::NTuple{Nv, UniformMesh{T}} : tuple of velocity meshes (one per velocity dimension)

t_interp_sp::NTuple{Nsp, InterpolationType{T}} : tuple of space interpolations (one per space dimension)

t_interp_v::NTuple{Nv, InterpolationType{T}} : tuple of velocity interpolations(one per velocity dimension)

dt_base::T : time delta for one advection series

Keywords

tab_coef=[1//2, 1//1, 1//2] : coefficient table for one advection series, the coefficients at odd indexes is for space advection series, the coefficients at even indexes is for velocity advection series

tab_fct=[identity, identity, identity] : function table for one advection series, with the same indexes than tab_coef

Implementation

sizeall : tuple of the sizes of all dimensions (space before velocity)

sizeitr : tuple of iterators of indexes of each dimension

t_mesh_sp : tuple of space meshes

t_mesh_v : tuple of velocity meshes

t_interp_sp : tuple of space interpolation types

t_interp_v : tuple of velocity interpolation types

dt_base::T : time unit of an advection series

tab_coef : coefficient table

tab_fct : function table

v_square : precompute for ke

nbsplit : number of slices for split

mpiid : MPI id

Throws

ArgumentError : Nsp must be less or equal to Nv.

SemiLagrangian.AdvectionData — Type

AdvectionData{T,Nsp,Nv,Nsum,timeopt}
AdvectionData(
adv::Advection{T,Nsp,Nv,Nsum,timeopt}, 
data::Array{T,Nsum},
parext)

Mutable structure that contains variable parameters of advection series

Type parameters

T::DataType : type of data

Nsp : number of space dimensions

Nv : number of velocity dimensions

Nsum : the total number of dimensions (Nsum = Nsp + Nv)

timeopt::TimeOptimization : time optimization

Arguments

adv::Advection{T,Nsp,Nv,Nsum} : link to the constant data of this advection

data::Array{T,Nsum} : Initial data of this advection

parext : external data of this advection to compute alpha of each interpolations

Implementation

adv::Advection{T,Nsp,Nv,Nsum,timeopt} : link to the constant data of this advection

state_coef::Int : state that is the index of tab_coef, it is from one to lenth(tab_coef)

state_dim::Int : the dimension index, from 1 to Nsp in space states, from one to Nv in velocity state

data:Array{T,Nsum} : it is the working buffer

bufdata::Vector{T} : vector of the same size of the working buffer

t_buf::NTuple{Nsum, Array{T,2}} : tuple of buffer that is used to get the linear data for interpolation, one buffer per thread

cache_alpha::Union{T,Nothing} : cache for precal, the precal is compute only when the alpha or decint values change

cache_decint::Int64 : for precal cache

cache_precal::Vector{T} : for precal cache

parext::ExtDataAdv : external data of this advection to compute alpha of each interpolations

Methods to define

initcoef!(parext::AbstractExtDataAdv, self::AdvectionData) : this method called at the beginning of each advection to initialize parext data. The self.parext mutable structure is the only data that init! can modify otherwise it leads to unpredictable behaviour.

getalpha(parext::AbstractExtDataAdv, self::AdvectionData, ind) : return the alpha number that is used for interpolation.

getperm(parext::AbstractExtDataAdv, advd::AdvectionData) : get the permutation of the dimension as a function of the current state

SemiLagrangian.B_Spline — Type

B_Spline{T,iscirc, order} <: InterpolationType{T,iscirc, order}

Abstract supertype for all bspline interpolation type

SemiLagrangian.InterpolationType — Type

InterpolationType{T, iscirc, order}

Abstract supertype for all interpolation type

SemiLagrangian.Lagrange — Type

Lagrange{T, iscirc, order, N} <: InterpolationType{T, iscirc, order}
Lagrange(order, T::DataType=Float64; iscirc::Bool=true)

Type containing Lagrange Polynomials coefficients for Lagrange interpolation

Type parameters

T : the type of data that is interpolate

iscirc::Bool : true if function is circular

order::Int: order of lagrange interpolation

N : type of integer, in fact Int64 or BigInt that is used to store lagrange polynomial

Implementation :

fact_order::N : factorial of the order

lagpol:Vector{Polynomial{N}} : vector of all lagrange polynomial, per example the k-th Lagrange polynomial for the designed order is lagpol[k+1]/fact_order

Arguments :

order : the order of interpolation

[T::DataType=Float64] : The type values to interpolate

SemiLagrangian.LuSpline — Type

struct LuSpline{T}

Structure of a LU decomposition of circular banded matrix, a LU decomposition can be stored in a Matrix which is equal to L + U - I. For a circular Banded matrix all non zero coefficients are in the band and in the last columns and lines

Implementation

band::Matrix{T} : matrix of size (kl+ku+1, n-kl)

ku : size of the band upper the diagonal

kl : size of the band lower the diagonal

iscirc : true if and only if original matrix is circular

isLU : true if LU decomposition has been perform

lastcols : only in circular case, Matrix of size (n-ku, kl) that represents teh last columns of matrix

lastrows : only in circular case, Matrix of size (n, ku) that represents the last rows of matrix

SemiLagrangian.PoissonConst — Type

PoissonConst{T, Nsp, Nv}
PoissonConst(adv::Advection{T, Nsp, Nv, Nsum}; isfftbig=true)

Constant data for the computation of poisson coefficients

Arguments

adv::Advection{T, Nsp, Nv, Nsum, timeopt} : Advection constant data

isfftbig=true: if true compute the fttbig structure

Implementation

adv : Advection constant data

`v_k' : vector of vector of fourier coefficents for integration for each space dimension

fctv_k : Array of space dimensions of the inverse of the norm of fourier coefficients

pfftbig : Fourier data for space dimensions

SemiLagrangian.PoissonVar — Type

PoissonVar{T, Nsp, Nv} <: AbstractExtDataAdv
PoissonVar(pc::PoissonConst{T, Nsp, Nv})

mutable structure of variable data for the poisson computation

Arguments

pc::PoissonConst{T, Nsp, Nv} : poisson constant data

Implementation

pc::PoissonConst{T, Nsp, Nv} : poisson constant data

rho::Array{T, Nsp} : result of the compute_charge that is the sum along velocity dimensions

t_elfield::NTuple{Nsp,Array{Complex{T}, Nsp}} : electric fields initialized at each beginning of velocity advection subseries

SemiLagrangian.UniformMesh — Type

UniformMesh(start, stop, length)

1D uniform mesh data.

length : number of points length-1 : number of cells

If you want remove the last point for periodic domain, set endpoint=false

Arguments

- `step`  : size step
- `points`: Array with node positions
- `width` : Distance between left and right edges.