Types

SemiLagrangian.B_Spline — Type

B_Spline{T, edge, order} <: AbstractInterpolation{T, edge, order}

Abstract supertype for all bspline interpolation type

SemiLagrangian.LuSpline — Type

struct LuSpline{T}
LuSpline(n, t::Vector{T}; iscirc=true, isLU=true) where{T}

Structure of a LU decomposition of circular banded matrix, a LU decomposition can be stored in a Matrix which is equal to L + U - I. For a circular Banded matrix all non zero coefficients are in the band and in the last columns and lines

Implementation

band::Matrix{T} : matrix of size (kl+ku+1, n-kl)

ku : size of the band upper the diagonal

kl : size of the band lower the diagonal

iscirc : true if and only if original matrix is circular

isLU : true if LU decomposition has been perform

lastcols : only in circular case, Matrix of size (n-ku, kl) that represents teh last columns of matrix

lastrows : only in circular case, Matrix of size (n, ku) that represents the last rows of matrix

Arguments

'n' : size of the matrix

't::Vector{T}` : vector of all values, the size is order+1, where order is the order of the spline.

SemiLagrangian.PoissonConst — Type

PoissonConst{T, Nsp, Nv}
PoissonConst(adv::Advection{T, Nsp, Nv, Nsum}; isfftbig=true)

Constant data for the computation of poisson coefficients

Arguments

adv::Advection{T, Nsp, Nv, Nsum, timeopt} : Advection constant data

isfftbig=true: if true compute the fttbig structure

Implementation

adv : Advection constant data

`v_k' : vector of vector of fourier coefficents for integration for each space dimension

fctv_k : Array of space dimensions of the inverse of the norm of fourier coefficients

pfftbig : Fourier data for space dimensions

SemiLagrangian.PoissonVar — Type

PoissonVar{T, Nsp, Nv, Nsum} <: AbstractExtDataAdv
PoissonVar(pc::PoissonConst{T, Nsp, Nv})

mutable structure of variable data for the poisson computation

Arguments

pc::PoissonConst{T, Nsp, Nv} : poisson constant data

Implementation

pc::PoissonConst{T, Nsp, Nv} : poisson constant data

rho::Array{T, Nsp} : result of the compute_charge that is the sum along velocity dimensions

t_elfield::NTuple{Nsp,Array{Complex{T}, Nsp}} : electric fields initialized at each beginning of velocity advection subseries